+++ lib/oxidized/model/procurve.rb
@@ -79,6 +79,13 @@ class Procurve < Oxidized::Model
   end
 
   cfg :telnet, :ssh do
+    # preferred way to handle additional passwords
+    if vars :enable
+      post_login do
+        send "enable\n"
+        cmd vars(:enable)
+      end
+    end
     post_login 'no page'
     pre_logout "logout\ny\nn"
   end